Added in some features
Asks for path to log, supports gzip compressed logfiles, also finds
users IP and geolocates them and builds the gmap to their coordinates,
fixed a mistake in the private IP checker thanks to Terry IP for
pointing that out and for bringing up the compressed log support issue.
I thought this was enough change to illicit a change in version and up
it to 0.2a.